Organized the code of the check algorithm part of the default module.#182
